+++ src/hash_table.c
@@ -329,7 +329,7 @@ hash_table_lookup_int32(hash_table_t * h, const char *key, int32 *val)
     if (rv != 0)
         return rv;
     if (val)
-        *val = (int32)vval;
+        *val = (int32)(size_t)vval;
     return 0;
 }
 
@@ -365,7 +365,7 @@ hash_table_lookup_bkey_int32(hash_table_t * h, const char *key, size_t len, int3
     if (rv != 0)
         return rv;
     if (val)
-        *val = (int32)vval;
+        *val = (int32)(size_t)vval;
     return 0;
 }
